Fix error: conversion from ‘std::streamoff’ {aka ‘long long int’} to ‘std::__cxx11::basic_string<char>::size_type’ {aka ‘unsigned int’} may change value [-Werror=conversion]

+++ src/mini/ini.h
@@ -340,7 +340,7 @@ namespace mINI
 		{
 			std::string fileContents;
 			fileReadStream.seekg(0, std::ios::end);
-			fileContents.resize(fileReadStream.tellg());
+			fileContents.resize(static_cast<std::size_t>(fileReadStream.tellg()));
 			fileReadStream.seekg(0, std::ios::beg);
 			std::size_t fileSize = fileContents.size();
 			fileReadStream.read(&fileContents[0], fileSize);
